Add dependencies, wallet_router, changing schemas, changing wallet service to make it correctly work, change 'Integer' to 'BigInteger' to these valuse 'bandwidth','energy'

+++ app/models/models.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import declarative_base
 from sqlalchemy.sql.schema import MetaData
-from sqlalchemy import Column, Integer, String, Float, DateTime, func
+from sqlalchemy import Column, Integer, String, Float, DateTime, func, BigInteger
 
 models_metadata = MetaData()
 Base = declarative_base(metadata=models_metadata)
@@ -11,7 +11,7 @@ class TronWallet(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     wallet_address = Column(String(256), nullable=False)
-    bandwidth = Column(Integer, nullable=False)
-    energy = Column(Integer, nullable=False)
+    bandwidth = Column(BigInteger, nullable=False)
+    energy = Column(BigInteger, nullable=False)
     balance_trx = Column(Float, nullable=False)
     created_at = Column(DateTime, default=func.now())
